+++ app/src/test/kotlin/com/arflix/tv/data/repository/IptvVodSearchPreferenceTest.kt
@@ -3,6 +3,7 @@ package com.arflix.tv.data.repository
 import com.arflix.tv.ui.screens.settings.SettingsUiState
 import com.arflix.tv.util.IPTV_VOD_SEARCH_ENABLED_KEY
 import com.google.common.truth.Truth.assertThat
+import com.google.gson.Gson
 import org.junit.Test
 
 class IptvVodSearchPreferenceTest {
@@ -15,5 +16,19 @@ class IptvVodSearchPreferenceTest {
     @Test
     fun settingsUiStateDefaultsVodSearchToEnabled() {
         assertThat(SettingsUiState().iptvVodSearchEnabled).isTrue()
+    }
+
+    @Test
+    fun legacyCloudStateLeavesVodSearchUnsetForEnabledFallback() {
+        val state = Gson().fromJson("{}", IptvCloudProfileState::class.java)
+
+        assertThat(state.vodSearchEnabled).isNull()
+    }
+
+    @Test
+    fun cloudStatePreservesDisabledVodSearch() {
+        val state = Gson().fromJson("{\"vodSearchEnabled\":false}", IptvCloudProfileState::class.java)
+
+        assertThat(state.vodSearchEnabled).isFalse()
     }
 }